requirements.txt
streamlit>=1.44,<2
jinja2>=3.1,<4
sqlparse>=0.5,<1
PyYAML>=6.0,<7
pyperclip>=1.9,<2
pytest>=8.0,<9
